requirements.txt
pandas 
duckdb 
requests 
requests-html 
pyarrow 
papermill
snowflake-connector-python
apache-airflow-providers-papermill
deltalake